fixed transaction

+++ src/main/java/pl/robakowski/transactions/Transaction.java
@@ -2,6 +2,8 @@ package pl.robakowski.transactions;
 
 import com.dslplatform.json.CompiledJson;
 
+import java.math.BigDecimal;
+
 @CompiledJson
-public record Transaction(String debitAccount, String creditAccount, float amount) {
+public record Transaction(String debitAccount, String creditAccount, BigDecimal amount) {
 }
